requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
apscheduler==3.10.4
sqlalchemy==2.0.30
pydantic==2.7.4
xgboost==2.0.3
fastf1==3.8.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
pyyaml==6.0.1
pyarrow==16.1.0
pytest==8.2.2
# redis==5.0.4
